requirements.txt
anyio @ file:///Users/runner/miniforge3/conda-bld/anyio_1634574651279/work/dist
appnope @ file:///Users/runner/miniforge3/conda-bld/appnope_1635819655903/work
argon2-cffi @ file:///Users/runner/miniforge3/conda-bld/argon2-cffi_1633990662093/work
async-generator==1.10
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1620387926260/work
Babel @ file:///home/conda/feedstock_root/build_artifacts/babel_1619719576210/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1629908509068/work
brotlipy==0.7.0
certifi==2021.10.8
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1635832827049/work
chardet @ file:///Users/runner/miniforge3/conda-bld/chardet_1635814945668/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1626371162869/work
click==7.1.2
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1635145350995/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
dataclasses @ file:///home/conda/feedstock_root/build_artifacts/dataclasses_1628958434797/work
debugpy @ file:///Users/runner/miniforge3/conda-bld/debugpy_1635844775611/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1631346842025/work
defusedxml @ file:///home/conda/feedstock_root/build_artifacts/defusedxml_1615232257335/work
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
Flask @ file:///home/conda/feedstock_root/build_artifacts/flask_1633384704162/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1609836280497/work
importlib-metadata @ file:///Users/runner/miniforge3/conda-bld/importlib-metadata_1635818343420/work
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1634726670251/work/dist/ipykernel-6.4.2-py3-none-any.whl
ipython @ file:///Users/runner/miniforge3/conda-bld/ipython_1635632578382/work
ipython-genutils==0.2.0
itsdangerous @ file:///home/conda/feedstock_root/build_artifacts/itsdangerous_1621360792037/work
jedi @ file:///Users/runner/miniforge3/conda-bld/jedi_1635824121556/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1633656206378/work
json5 @ file:///home/conda/feedstock_root/build_artifacts/json5_1600692310011/work
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema_1634752161479/work
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1633454794268/work
jupyter-core @ file:///Users/runner/miniforge3/conda-bld/jupyter_core_1635407494759/work
jupyter-server @ file:///home/conda/feedstock_root/build_artifacts/jupyter_server_1635811281209/work
jupyterlab @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_1634741878908/work
jupyterlab-pygments @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_pygments_1601375948261/work
jupyterlab-server @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_server_1632590716858/work
kiwisolver @ file:///Users/runner/miniforge3/conda-bld/kiwisolver_1635836812195/work
MarkupSafe @ file:///opt/concourse/worker/volumes/live/1bcf0940-5bf8-4171-61e3-2133f5885e8b/volume/markupsafe_1621528148241/work
matplotlib @ file:///Users/runner/miniforge3/conda-bld/matplotlib-suite_1632416605047/work
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
mistune @ file:///Users/runner/miniforge3/conda-bld/mistune_1635844789148/work
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
nbclassic @ file:///home/conda/feedstock_root/build_artifacts/nbclassic_1635436909771/work
nbclient @ file:///home/conda/feedstock_root/build_artifacts/nbclient_1629120697898/work
nbconvert @ file:///Users/runner/miniforge3/conda-bld/nbconvert_1632536029521/work
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1617383142101/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1617163391303/work
notebook @ file:///home/conda/feedstock_root/build_artifacts/notebook_1634668519289/work
numpy @ file:///opt/concourse/worker/volumes/live/21fb9bb9-9731-4901-4403-2bc27cc2cc33/volume/numpy_and_numpy_base_1634106739792/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1625323647219/work
pandocfilters @ file:///home/conda/feedstock_root/build_artifacts/pandocfilters_1631603243851/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1617148930513/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1630696651849/work
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1635538335951/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1634921890410/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1629119114968/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1633192417276/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1635593541381/work
pyrsistent @ file:///Users/runner/miniforge3/conda-bld/pyrsistent_1635836928927/work
PySocks @ file:///Users/runner/miniforge3/conda-bld/pysocks_1610291470465/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytorch-pretrained-vit==0.0.7
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1633452062248/work
pyzmq @ file:///Users/runner/miniforge3/conda-bld/pyzmq_1631793531263/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1626393743643/work
Send2Trash @ file:///home/conda/feedstock_root/build_artifacts/send2trash_1628511208346/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio @ file:///Users/runner/miniforge3/conda-bld/sniffio_1635845059573/work
terminado @ file:///Users/runner/miniforge3/conda-bld/terminado_1631192978302/work
testpath @ file:///home/conda/feedstock_root/build_artifacts/testpath_1621261527237/work
torch==1.10.0
torchvision==0.11.1
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1635819729986/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1635260543454/work
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1632313171031/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1632350318291/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
webencodings==0.5.1
websocket-client @ file:///Users/runner/miniforge3/conda-bld/websocket-client_1635860050933/work
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1621518206714/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1633302054558/work
